+++ docs/changelog.md
@@ -4,6 +4,7 @@
 
 - Basic support for checking `Final` and for checking re-assignments
   to variables declared with a specific type (#505)
+- Correctly check the `self` argument to `@property` getters (#506)
 - Correctly track assignments of variables inside `try` blocks
   and inside `with` blocks that may suppress exceptions (#504)
 - Support mappings that do not inherit from `collections.abc.Mapping`